PIC-104: Слияние изменений
