--- src/components/Emojies/Emojies.hbs
@@ -1,11 +1,0 @@
-<div class="emojies">
-    {{#each emojies}}
-    <div class="emojies__row">
-        {{#each this}}
-        <div class="emojies__column pointer-icon">
-            {{{this}}}
-        </div>
-        {{/each}}
-    </div>
-    {{/each}}
-</div>
