+++ src/components/Emojis/Emojis.hbs
@@ -1,0 +1,11 @@
+<div class="emojis">
+    {{#each emojis}}
+    <div class="emojis__row">
+        {{#each this}}
+        <div class="emojis__column pointer-icon">
+            {{{this}}}
+        </div>
+        {{/each}}
+    </div>
+    {{/each}}
+</div>
