--- src/consts/emojies.ts
@@ -1,10 +1,0 @@
-const EmojiesList = {
-    Emojie_0: '😍',
-    Emojie_1: '😏',
-    Emojie_2: '😡',
-    Emojie_3: '💩',
-    Emojie_4: '👎',
-    Emojie_5: '❤️'
-};
-
-export default EmojiesList;
