+++ src/consts/emojis.ts
@@ -1,0 +1,10 @@
+const EmojisList = {
+    Emoji_0: '😍',
+    Emoji_1: '😏',
+    Emoji_2: '😡',
+    Emoji_3: '💩',
+    Emoji_4: '👎',
+    Emoji_5: '❤️'
+};
+
+export default EmojisList;
